+++ apps/achievements-remote/src/app/achievements/data-access/index.ts
@@ -1,3 +1,4 @@
 export * from './achievement.service';
 export * from './achievements.facade';
 export * from './mini-bus.service';
+export * from './achievement-service.interface';