+++ jest.preset.js
@@ -2,6 +2,6 @@ const nxPreset = require('@nx/jest/preset').default;
 
 module.exports = {
   ...nxPreset,
-  transformIgnorePatterns: ['node_modules/(?!.*\\.mjs$|@jsverse)'],
+  transformIgnorePatterns: [String.raw`node_modules/(?!.*\.mjs$|@jsverse)`],
   coverageReporters: ['html', 'lcov', 'text-summary'],
 };